+++ CHANGELOG.md
@@ -23,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   in memory and when written as a file
 - added new subgrid type `LagrangeSubgridV2`, which allows for different `x1`
   and `x2` limits, for example for DIS processes
-- added new switches to the subcommand `info`: `--get`, `--show`,
+- added new switches to the subcommand `info`: `--get`, `--keys`, `--show`,
   which lets one read a single or all key-values pairs of each grid file.
 - added new subcommand `set` which allows to modify the key-value storage